requirements.txt
Django==1.10.1
Pillow==3.3.1
django-ckeditor==5.1.1
django-cleanup==0.4.2
django-embed-video==1.1.0
psycopg2==2.6.2
requests==2.11.1
